Created a common web/server/test/conftest Move there all standard dirs fixtures
