+++ services/web/server/tests/conftest.py
@@ -1,0 +1,70 @@
+""" Main test configuration
+
+    EXPECTED: simcore_service_webserver installed
+
+"""
+# pylint: disable=unused-argument
+# pylint: disable=bare-except
+# pylint:disable=redefined-outer-name
+
+import logging
+import sys
+from pathlib import Path
+
+import pytest
+
+import simcore_service_webserver
+
+current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
+log = logging.getLogger(__name__)
+
+# mute noisy loggers
+logging.getLogger("openapi_spec_validator").setLevel(logging.WARNING)
+logging.getLogger("sqlalchemy").setLevel(logging.WARNING)
+
+## HELPERS
+sys.path.append(str(current_dir / 'helpers'))
+
+
+## FIXTURES: standard paths
+
+@pytest.fixture(scope='session')
+def package_dir() -> Path:
+    """ osparc-simcore installed directory """
+    dirpath = Path(simcore_service_webserver.__file__).resolve().parent
+    assert dirpath.exists()
+    return dirpath
+
+@pytest.fixture(scope='session')
+def osparc_simcore_root_dir() -> Path:
+    """ osparc-simcore repo root dir """
+    WILDCARD = "services/web/server"
+
+    root_dir = Path(current_dir)
+    while not any(root_dir.glob(WILDCARD)) and root_dir != Path("/"):
+        root_dir = root_dir.parent
+
+    msg = f"'{root_dir}' does not look like the git root directory of osparc-simcore"
+    assert root_dir.exists(), msg
+    assert any(root_dir.glob(WILDCARD)), msg
+    assert  any(root_dir.glob(".git")), msg
+
+    return root_dir
+
+@pytest.fixture(scope="session")
+def env_devel_file(osparc_simcore_root_dir) -> Path:
+    env_devel_fpath = osparc_simcore_root_dir / ".env-devel"
+    assert env_devel_fpath.exists()
+    return env_devel_fpath
+
+@pytest.fixture(scope='session')
+def api_specs_dir(osparc_simcore_root_dir: Path) -> Path:
+    specs_dir = osparc_simcore_root_dir/ "api" / "specs" / "webserver"
+    assert specs_dir.exists()
+    return specs_dir
+
+@pytest.fixture(scope='session')
+def fake_data_dir() -> Path:
+    dirpath = (current_dir / "data").resolve()
+    assert dirpath.exists()
+    return dirpath
